+++ waf-openapi-20211001/src/Models/DescribeProductInstancesResponseBody/productInstances/resourcePorts/certificates.php
@@ -11,15 +11,27 @@ class certificates extends Model
     /**
      * @var string
      */
+    public $appliedType;
+
+    /**
+     * @var string
+     */
     public $certificateId;
 
     /**
      * @var string
      */
     public $certificateName;
+
+    /**
+     * @var string
+     */
+    public $domain;
     protected $_name = [
+        'appliedType' => 'AppliedType',
         'certificateId' => 'CertificateId',
         'certificateName' => 'CertificateName',
+        'domain' => 'Domain',
     ];
 
     public function validate()
@@ -30,12 +42,20 @@ class certificates extends Model
     public function toArray($noStream = false)
     {
         $res = [];
+        if (null !== $this->appliedType) {
+            $res['AppliedType'] = $this->appliedType;
+        }
+
         if (null !== $this->certificateId) {
             $res['CertificateId'] = $this->certificateId;
         }
 
         if (null !== $this->certificateName) {
             $res['CertificateName'] = $this->certificateName;
+        }
+
+        if (null !== $this->domain) {
+            $res['Domain'] = $this->domain;
         }
 
         return $res;
@@ -49,12 +69,20 @@ class certificates extends Model
     public static function fromMap($map = [])
     {
         $model = new self();
+        if (isset($map['AppliedType'])) {
+            $model->appliedType = $map['AppliedType'];
+        }
+
         if (isset($map['CertificateId'])) {
             $model->certificateId = $map['CertificateId'];
         }
 
         if (isset($map['CertificateName'])) {
             $model->certificateName = $map['CertificateName'];
+        }
+
+        if (isset($map['Domain'])) {
+            $model->domain = $map['Domain'];
         }
 
         return $model;